requirements.txt
numpy
scipy
astropy
photutils
pandas
pytest>=3.6
tqdm
matplotlib
pillow
panoptes-client
coverage
python-coveralls
pytest-cov
codecov
requests
numba
../shared-astro-utilities  # local install, TODO move to github